backend/requirements.txt
alembic==1.13.2
factory-boy==3.3.0
fastapi==0.111.1
mangum==0.17.0
psycopg2-binary==2.9.9
pydantic[email]
pydantic-settings==2.3.4
pytest==8.2.2
python-json-logger==2.0.7
sqlmodel==0.0.20
uvicorn==0.30.3